requirements.txt
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
firebase==3.0.1
firebase-admin==4.3.0
Flask==1.1.2
gcloud==0.17.0
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
msgpack==1.0.0
natsort==7.0.1
numpy==1.19.1
oauth2client==3.0.0
opencv-python==4.4.0.42
pandas==1.1.1
passlib==1.7.2
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyfcm==1.4.7
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.7.0
rsa==4.6
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
